fix(ci): restore monitor engine hotspot budgets
