+++ native/rust/crates/ripdpi-monitor-engine/src/session/lifecycle.rs
@@ -1,20 +1,19 @@
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
 use std::thread::JoinHandle;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use rustls::client::danger::ServerCertVerifier;
 
 use crate::execution::UnavailableCandidateRuntimeLauncher;
 use crate::platform::NoopMonitorPlatformBridge;
-use crate::types::{EngineScanRequestWire, SharedState};
+use crate::types::SharedState;
 use crate::{CandidateRuntimeLauncher, MonitorPlatformBridge};
 
-use super::log_level::parse_native_log_level;
-use super::reaper::WORKER_REAPER;
-use super::validation::ValidatedScanRequest;
-use super::wire_json::{passive_events_to_json, progress_to_json, report_to_json};
-use super::worker::{ScanWorkerConfig, join_finished_worker_locked, spawn_scan_worker};
+mod cancellation;
+mod reporting;
+mod start;
+mod teardown;
 
 #[derive(Default)]
 struct ScanControl {
@@ -91,147 +90,6 @@ impl MonitorSession {
             platform_bridge,
             candidate_runtime_launcher,
         }
-    }
-
-    pub fn start_scan(&self, session_id: String, request: EngineScanRequestWire) -> Result<(), String> {
-        if self.destroyed.load(Ordering::Acquire) {
-            return Err("diagnostics session destroyed".to_string());
-        }
-        let request = ValidatedScanRequest::try_from(request)?;
-        let native_log_level = parse_native_log_level(request.as_wire().native_log_level.as_deref())?;
-        let mut worker_guard = self.worker.lock().map_err(|_| "monitor worker poisoned".to_string())?;
-        join_finished_worker_locked(&mut worker_guard);
-        if worker_guard.is_some() || self.starting.load(Ordering::Acquire) {
-            return Err("diagnostics scan already running".to_string());
-        }
-        self.cancel.store(false, Ordering::Release);
-        *self.active_session_id.lock().map_err(|_| "monitor session id poisoned".to_string())? =
-            Some(session_id.clone());
-        let scan_deadline =
-            Instant::now() + Duration::from_millis(request.as_wire().scan_deadline_ms.unwrap_or(360_000));
-        *self.scan_control.lock().map_err(|_| "monitor scan control poisoned".to_string())? =
-            ScanControl { deadline: Some(scan_deadline), report_delivered: false };
-        *self.cancellation_reason.lock().map_err(|_| "monitor cancellation state poisoned".to_string())? = None;
-        {
-            let mut shared = self.shared.lock().map_err(|_| "monitor shared state poisoned".to_string())?;
-            shared.progress = None;
-            shared.report = None;
-            shared.checkpoint_report = None;
-            shared.log_context = request.as_wire().log_context.clone();
-        }
-        self.starting.store(true, Ordering::Release);
-        let _starting_guard = StartingGuard(&self.starting);
-        drop(worker_guard);
-        self.platform_bridge.clear_passive_events(&session_id);
-        if self.destroyed.load(Ordering::Acquire) {
-            return Err("diagnostics session destroyed".to_string());
-        }
-        let domain_request = request.into();
-        let worker_config = ScanWorkerConfig::new(
-            scan_deadline,
-            self.cancellation_reason.clone(),
-            self.tls_verifier.clone(),
-            self.platform_bridge.clone(),
-            self.candidate_runtime_launcher.clone(),
-            native_log_level,
-        );
-        let mut worker_guard = self.worker.lock().unwrap_or_else(std::sync::PoisonError::into_inner);
-        if self.destroyed.load(Ordering::Acquire) {
-            return Err("diagnostics session destroyed".to_string());
-        }
-        *worker_guard = Some(spawn_scan_worker(
-            self.shared.clone(),
-            self.cancel.clone(),
-            session_id,
-            domain_request,
-            worker_config,
-        ));
-        Ok(())
-    }
-
-    pub fn cancel_scan(&self) {
-        // Serialize cancellation with start_scan so initialization cannot clear a newly published request.
-        let _worker_guard = (!self.starting.load(Ordering::Acquire))
-            .then(|| self.worker.lock().unwrap_or_else(std::sync::PoisonError::into_inner));
-        let reason = self.scan_control.lock().ok().map(|scan_control| {
-            if scan_control.deadline.is_some_and(|deadline| Instant::now() >= deadline) {
-                crate::types::ScanTerminationReason::DeadlineExceeded
-            } else {
-                crate::types::ScanTerminationReason::UserCancelled
-            }
-        });
-        if let Ok(mut cancellation_reason) = self.cancellation_reason.lock()
-            && cancellation_reason.is_none()
-        {
-            *cancellation_reason = reason;
-        }
-        self.cancel.store(true, Ordering::Release);
-    }
-
-    pub fn poll_progress_json(&self) -> Result<Option<String>, String> {
-        let shared = self.shared.lock().map_err(|_| "monitor shared state poisoned".to_string())?;
-        progress_to_json(shared.progress.as_ref())
-    }
-
-    pub fn take_report_json(&self) -> Result<Option<String>, String> {
-        self.try_join_worker();
-        let mut scan_control = self.scan_control.lock().map_err(|_| "monitor scan control poisoned".to_string())?;
-        if scan_control.report_delivered {
-            return Ok(None);
-        }
-        // Ordering: observes cancellation published by cancel_scan before exposing a checkpoint.
-        let cancellation_requested = self.cancel.load(Ordering::Acquire);
-        let cancellation_reason =
-            self.cancellation_reason.lock().map_err(|_| "monitor cancellation state poisoned".to_string())?.clone();
-        let mut shared = self.shared.lock().map_err(|_| "monitor shared state poisoned".to_string())?;
-        let scan_finished = shared.progress.as_ref().is_none_or(|progress| progress.is_finished);
-        if !scan_finished && !cancellation_requested {
-            return Ok(None);
-        }
-        if cancellation_requested && !scan_finished {
-            let mut checkpoint = shared.checkpoint_report.take();
-            if let Some(report) = checkpoint.as_mut() {
-                report.completion_kind = crate::types::ScanCompletionKind::PartialResults;
-                report.termination_reason = cancellation_reason;
-            }
-            let json = report_to_json(checkpoint.as_ref())?;
-            scan_control.report_delivered = json.is_some();
-            return Ok(json);
-        }
-        let report = shared.report.take();
-        let json = report_to_json(report.as_ref())?;
-        scan_control.report_delivered = json.is_some();
-        Ok(json)
-    }
-
-    pub fn poll_passive_events_json(&self) -> Result<Option<String>, String> {
-        let session_id = self.active_session_id.lock().map_err(|_| "monitor session id poisoned".to_string())?.clone();
-        let events = session_id.as_deref().map(|id| self.platform_bridge.drain_passive_events(id)).unwrap_or_default();
-        passive_events_to_json(events)
-    }
-
-    /// Cancel the active scan and retire its worker without blocking the caller.
-    ///
-    /// An unfinished worker is joined by the process-wide diagnostics reaper.
-    /// This keeps JNI teardown bounded even when a probe is inside blocking I/O;
-    /// the worker still owns its state until it exits and is reaped.
-    pub fn destroy(&self) {
-        self.destroyed.store(true, Ordering::Release);
-        self.cancel_scan();
-        let handle = self.worker.lock().ok().and_then(|mut worker_guard| worker_guard.take());
-        if let Some(handle) = handle
-            && let Err(handle) = WORKER_REAPER.reap(handle)
-        {
-            log::error!("detaching diagnostics worker because the bounded reaper is saturated or unavailable");
-            drop(handle);
-        }
-    }
-
-    fn try_join_worker(&self) {
-        let Ok(mut worker_guard) = self.worker.lock() else {
-            return;
-        };
-        join_finished_worker_locked(&mut worker_guard);
     }
 }
