+++ native/rust/crates/ripdpi-monitor-engine/src/session/worker.rs
@@ -6,11 +6,13 @@ use std::time::Instant;
 use log::LevelFilter;
 use rustls::client::danger::ServerCertVerifier;
 
-use crate::engine::run_engine_scan;
 use crate::types::{ScanRequest, SharedState};
 use crate::{CandidateRuntimeLauncher, MonitorPlatformBridge};
 
 use super::panic_state::record_panic_terminal_state;
+use run::run_scan;
+
+mod run;
 
 pub(super) struct ScanWorkerConfig {
     scan_deadline: Instant,
@@ -73,28 +75,6 @@ pub(super) fn join_finished_worker_locked(worker_guard: &mut Option<JoinHandle<(
             let _ = handle.join();
         }
     }
-}
-
-fn run_scan(
-    shared: Arc<Mutex<SharedState>>,
-    cancel: Arc<AtomicBool>,
-    session_id: String,
-    request: ScanRequest,
-    config: ScanWorkerConfig,
-) {
-    let _log_scope = config
-        .native_log_level
-        .map(|level| config.platform_bridge.scoped_log_level("diagnostics_native".to_string(), level));
-    run_engine_scan(
-        shared,
-        cancel,
-        session_id,
-        request,
-        config.scan_deadline,
-        config.cancellation_reason,
-        config.tls_verifier,
-        config.candidate_runtime_launcher,
-    );
 }
 
 #[cfg(test)]
